Fix for 7.0

+++ Model/PersonDetails.cs
@@ -6,13 +6,13 @@ namespace MiniMappingway.Model
     {
         public string Name { get; }
 
-        public uint Id { get; }
+        public ulong Id { get; }
 
         public string SourceName { get; }
 
         public IntPtr Ptr { get; }
 
-        public PersonDetails(string name, uint id, string sourceName, IntPtr ptr)
+        public PersonDetails(string name, ulong id, string sourceName, IntPtr ptr)
         {
             Name = name;
             Id = id;
